Move out of headless tests
